requirements.txt
openai>=1.30.0
pydantic>=2.6
python-dotenv>=1.0
markdown-it-py>=3.0
Pillow>=10.3
typer>=0.12
rich>=13.7
PyYAML>=6.0
pytest>=8.2